Removes index_sirsi_hourly.sh from catchup.sh.
If a full reindex of MARC data completes before 6am, the hourly file
from the previous day will be indexed after the nightly file that was
generated just before midnight. There's a chance that the data in the
hourly file is outdated versus the data in the nightly file. We should
rely on the cronjob to index the latest hourly file.