Simplify and and unify logic for when a link is considered full-text.
We were handling "blank" a indicator2 a bit differently between the marc_links_struct and the url_fulltext field (which is used to determine of Access = Online)

This unifies that approach and simplifies it a bit.

We are no longer trying to treat "blank" indicator2 as full-text anymore in the Online facet.  We only appeared to be in the code before in the marc_links_struct, however a nil or blank indicator2 comes through as " " (failing the .empty? check), so was always falling through as falsey for the full-text check anyway (the behavior we likely wanted).

According to https://www.loc.gov/marc/bibliographic/bd856.html, an indicator2 of 0 or 1 is either the resource itself or a version of the resource.  We then take those 856s and cast out any links that have labels that arouse suspicion about the full-text nature of the link (e.g. table of contents).  Any other indicators are treated as not full-text, which is the behavior of what we're showing in the UI. This should have the Online facet more accurately match where we will be displaying links to the user.