Create a helper method for reading one catkey from the db
This implements and documents (in the README) a helper method that
mimics functionality we were using in tests: fetching a single
record from the database by catkey, using our custom SQL query.

You can use it to see a record "as the indexer sees it" and also
save that record to a JSON file (e.g. for use with #816).