Ensure we disable old course reserves lookup during indexing
Because we load the sirsi config after overriding reserves_lookup,
we weren't actually turning off the lookup of old course reserve
data, so it was being used at index time to populate item_display
and other fields.

This changes the order so that we override the method after loading
the sirsi_config.