Ensure correct record id is set to enable deletion from geoblacklight
GeoBlacklight records don't have an id field like SearchWorks ones
do - instead they have the layer_slug_s field (in v1.0 schema), which
is set to the prefixed druid. This has to be set in the traject
context so that the indexer issues a delete using the id that the
record will actually have.

Fixes #597